Speed up php-fpm container on MacOS
The script setup_native_nfs_docker_osx.sh will activate the NFS enabling
to speed up the mounted storage.

Adding health check for php-fpm container, just running the command
`docker-compose ps` it will show if the container is healthiness.

The HEALTHCHECK instruction tells Docker how to test a container
to check that it is still working. This can detect cases such as a
web server stuck in an infinite loop and unable to handle new
connections, even though the server process is still running.

Separating some commands into RUN instructions allow the Docker to cache
the instruction result to reuse in another moment.
https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#leverage-build-cache

Co-authored-by: Kévin Dunglas <dunglas@gmail.com>
Co-authored-by: arti0090 <arti0090@gmail.com>